edit  chnages
edit  chnages

+++ application/controllers/Hospital.php
@@ -296,6 +296,87 @@ class Hospital extends CI_Controller {
 					$post=$this->input->post();
 					//echo '<pre>';print_r($_FILES);exit;
 						$hospital_details= $this->Hospital_model->get_hospital_details(base64_decode($post['hospital_id']));
+
+						if(isset($_FILES['kyc_file1']['name']) && $_FILES['kyc_file1']['name']!=''){
+							unlink("assets/kyc_documents/".$hospital_details['kyc_file1']);
+							$temp = explode(".", $_FILES["kyc_file1"]["name"]);
+							$file1 = round(microtime(true)) . '.' . end($temp);
+							move_uploaded_file($_FILES['kyc_file1']['tmp_name'], "assets/kyc_documents/" . $file1);
+						}else{
+							$file1='';
+						}
+						if(isset($_FILES['kyc_file2']['name']) && $_FILES['kyc_file2']['name']!=''){
+							unlink("assets/kyc_documents/".$hospital_details['kyc_file2']);
+							$temp = explode(".", $_FILES["kyc_file2"]["name"]);
+							$file2 =base64_decode($post['hospital_id']).round(microtime(true)) . '.' . end($temp);
+							move_uploaded_file($_FILES['kyc_file2']['tmp_name'], "assets/kyc_documents/" . $file2);
+						}else{
+							$file2='';
+						}
+						if(isset($_FILES['kyc_file3']['name']) && $_FILES['kyc_file3']['name']!=''){
+							unlink("assets/kyc_documents/".$hospital_details['kyc_file3']);
+							$temp = explode(".", $_FILES["kyc_file3"]["name"]);
+							$file3 =base64_decode($post['hospital_id']).'1'.round(microtime(true)) . '.' . end($temp);
+							move_uploaded_file($_FILES['kyc_file3']['tmp_name'], "assets/kyc_documents/" . $file3);
+						}else{
+							$file3='';
+						}
+					$fivedata=array(
+							'kyc_doc1'=>isset($post['kyc_doc1'])?$post['kyc_doc1']:'',
+							'kyc_doc2'=>isset($post['kyc_doc2'])?$post['kyc_doc2']:'',
+							'kyc_doc3'=>isset($post['kyc_doc3'])?$post['kyc_doc3']:'',
+							'kyc_file1'=>$file1,
+							'kyc_file2'=>$file2,
+							'kyc_file3'=>$file3,
+							'hos_updated_at'=>date('Y-m-d H:i:s')
+							);
+							$stepfive= $this->Hospital_model->update_hospital_details(base64_decode($post['hospital_id']),$fivedata);
+							if(count($stepfive)>0){
+								$this->session->set_flashdata('success',"Hospital Details are successfully Saved");
+								redirect('hospital');
+							}else{
+								$this->session->set_flashdata('error',"technical problem will occurred. Please try again.");
+								redirect('hospital/add/'.base64_encode(5).'/'.$post['hospital_id']);
+							}
+				}else{
+					$this->session->set_flashdata('error',"You have no permission to access");
+					redirect('dashboard');
+				}
+			
+		}else{
+			$this->session->set_flashdata('error','Please login to continue');
+			redirect('admin');
+		}
+	}
+	public function edit()
+	{
+		if($this->session->userdata('userdetails'))
+		{
+			if($admindetails['role_id']=1){
+					$data['tab']=base64_decode($this->uri->segment(4));
+					$data['hospital_id']=$this->uri->segment(3);
+					$hospital_id=base64_decode($this->uri->segment(3));
+					$data['hospital_details']= $this->Hospital_model->get_hospital_details($hospital_id);
+					//echo '<pre>';print_r($data);exit;
+					$this->load->view('admin/edithospital',$data);
+					$this->load->view('html/footer');
+			}else{
+					$this->session->set_flashdata('error',"You have no permission to access");
+					redirect('dashboard');
+			}
+		}else{
+			$this->session->set_flashdata('error','Please login to continue');
+			redirect('admin');
+		}
+	}
+	public function editpost(){
+		if($this->session->userdata('userdetails'))
+		{
+				if($admindetails['role_id']=1){
+					
+					$post=$this->input->post();
+					echo '<pre>';print_r($post);
+						$hospital_details= $this->Hospital_model->get_hospital_details(base64_decode($post['hospital_id']));
 						if(isset($_FILES['hos_bas_document']['name']) && $_FILES['hos_bas_document']['name']!=''){
 							$hospital_details= $this->Hospital_model->get_hospital_details(base64_decode($post['hospital_id']));
 							unlink("assets/hospital_basic_documents/".$hospital_details['hos_bas_document']);
@@ -338,7 +419,7 @@ class Hospital extends CI_Controller {
 						}else{
 							$file3=$hospital_details['kyc_file3'];
 						}
-					$fivedata=array(
+					$editdata=array(
 							'hos_con_number'=>isset($post['hos_con_number'])?$post['hos_con_number']:$hospital_details['hos_con_number'],
 							'hos_email_id'=>isset($post['hos_email_id'])?$post['hos_email_id']:$hospital_details['hos_email_id'],
 							'hos_representative'=>isset($post['hos_representative'])?$post['hos_representative']:$hospital_details['hos_representative'],
@@ -358,97 +439,28 @@ class Hospital extends CI_Controller {
 							'hos_bas_email'=>isset($post['hos_bas_email'])?$post['hos_bas_email']:$hospital_details['hos_bas_email'],
 							'hos_bas_nationali_id'=>isset($post['hos_bas_nationali_id'])?$post['hos_bas_nationali_id']:$hospital_details['hos_bas_nationali_id'],
 							'hos_bas_add1'=>isset($post['hos_bas_add1'])?$post['hos_bas_add1']:$hospital_details['hos_bas_add1'],
-							'kyc_doc1'=>isset($post['kyc_doc1'])?$post['kyc_doc1']:'',
-							'kyc_doc2'=>isset($post['kyc_doc2'])?$post['kyc_doc2']:'',
-							'kyc_doc3'=>isset($post['kyc_doc3'])?$post['kyc_doc3']:'',
+							'hos_bas_add2'=>isset($post['hos_bas_add2'])?$post['hos_bas_add2']:$hospital_details['hos_bas_add2'],
+							'hos_bas_zipcode'=>isset($post['hos_bas_zipcode'])?$post['hos_bas_zipcode']:$hospital_details['hos_bas_zipcode'],
+							'hos_bas_city'=>isset($post['hos_bas_city'])?$post['hos_bas_city']:$hospital_details['hos_bas_city'],
+							'hos_bas_state'=>isset($post['hos_bas_state'])?$post['hos_bas_state']:$hospital_details['hos_bas_state'],
+							'hos_bas_country'=>isset($post['hos_bas_country'])?$post['hos_bas_country']:$hospital_details['hos_bas_country'],
+							'hos_bas_document'=>$hos_bas_document,
+							'bank_holder_name'=>isset($post['bank_holder_name'])?$post['bank_holder_name']:$hospital_details['bank_holder_name'],
+							'bank_acc_no'=>isset($post['bank_acc_no'])?$post['bank_acc_no']:$hospital_details['bank_acc_no'],
+							'bank_name'=>isset($post['bank_name'])?$post['bank_name']:$hospital_details['bank_name'],
+							'bank_ifsc'=>isset($post['bank_ifsc'])?$post['bank_ifsc']:$hospital_details['bank_ifsc'],
+							'bank_document'=>$bank_document,
+							'kyc_doc1'=>isset($post['kyc_doc1'])?$post['kyc_doc1']:$hospital_details['kyc_doc1'],
+							'kyc_doc2'=>isset($post['kyc_doc2'])?$post['kyc_doc2']:$hospital_details['kyc_doc2'],
+							'kyc_doc3'=>isset($post['kyc_doc3'])?$post['kyc_doc3']:$hospital_details['kyc_doc3'],
 							'kyc_file1'=>$file1,
 							'kyc_file2'=>$file2,
 							'kyc_file3'=>$file3,
 							'hos_updated_at'=>date('Y-m-d H:i:s')
 							);
-							$stepfive= $this->Hospital_model->update_hospital_details(base64_decode($post['hospital_id']),$fivedata);
-							if(count($stepfive)>0){
-								$this->session->set_flashdata('success',"Hospital Details are successfully Saved");
-								redirect('hospital');
-							}else{
-								$this->session->set_flashdata('error',"technical problem will occurred. Please try again.");
-								redirect('hospital/add/'.base64_encode(5).'/'.$post['hospital_id']);
-							}
-				}else{
-					$this->session->set_flashdata('error',"You have no permission to access");
-					redirect('dashboard');
-				}
-			
-		}else{
-			$this->session->set_flashdata('error','Please login to continue');
-			redirect('admin');
-		}
-	}
-	public function edit()
-	{
-		if($this->session->userdata('userdetails'))
-		{
-			if($admindetails['role_id']=1){
-					$data['tab']=base64_decode($this->uri->segment(4));
-					$data['hospital_id']=base64_decode($this->uri->segment(3));
-					$hospital_id=base64_decode($this->uri->segment(3));
-					$data['hospital_details']= $this->Hospital_model->get_hospital_details($hospital_id);
-					//echo '<pre>';print_r($data);exit;
-					$this->load->view('admin/edithospital',$data);
-					$this->load->view('html/footer');
-			}else{
-					$this->session->set_flashdata('error',"You have no permission to access");
-					redirect('dashboard');
-			}
-		}else{
-			$this->session->set_flashdata('error','Please login to continue');
-			redirect('admin');
-		}
-	}
-	public function editpost(){
-		if($this->session->userdata('userdetails'))
-		{
-				if($admindetails['role_id']=1){
-					
-					$post=$this->input->post();
-					echo '<pre>';print_r($_FILES);exit;
-						$hospital_details= $this->Hospital_model->get_hospital_details(base64_decode($post['hospital_id']));
-
-						if(isset($_FILES['kyc_file1']['name']) && $_FILES['kyc_file1']['name']!=''){
-							unlink("assets/kyc_documents/".$hospital_details['kyc_file1']);
-							$temp = explode(".", $_FILES["kyc_file1"]["name"]);
-							$file1 = round(microtime(true)) . '.' . end($temp);
-							move_uploaded_file($_FILES['kyc_file1']['tmp_name'], "assets/kyc_documents/" . $file1);
-						}else{
-							$file1='';
-						}
-						if(isset($_FILES['kyc_file2']['name']) && $_FILES['kyc_file2']['name']!=''){
-							unlink("assets/kyc_documents/".$hospital_details['kyc_file2']);
-							$temp = explode(".", $_FILES["kyc_file2"]["name"]);
-							$file2 =base64_decode($post['hospital_id']).round(microtime(true)) . '.' . end($temp);
-							move_uploaded_file($_FILES['kyc_file2']['tmp_name'], "assets/kyc_documents/" . $file2);
-						}else{
-							$file2='';
-						}
-						if(isset($_FILES['kyc_file3']['name']) && $_FILES['kyc_file3']['name']!=''){
-							unlink("assets/kyc_documents/".$hospital_details['kyc_file3']);
-							$temp = explode(".", $_FILES["kyc_file3"]["name"]);
-							$file3 =base64_decode($post['hospital_id']).'1'.round(microtime(true)) . '.' . end($temp);
-							move_uploaded_file($_FILES['kyc_file3']['tmp_name'], "assets/kyc_documents/" . $file3);
-						}else{
-							$file3='';
-						}
-					$fivedata=array(
-							'kyc_doc1'=>isset($post['kyc_doc1'])?$post['kyc_doc1']:'',
-							'kyc_doc2'=>isset($post['kyc_doc2'])?$post['kyc_doc2']:'',
-							'kyc_doc3'=>isset($post['kyc_doc3'])?$post['kyc_doc3']:'',
-							'kyc_file1'=>$file1,
-							'kyc_file2'=>$file2,
-							'kyc_file3'=>$file3,
-							'hos_updated_at'=>date('Y-m-d H:i:s')
-							);
-							$stepfive= $this->Hospital_model->update_hospital_details(base64_decode($post['hospital_id']),$fivedata);
-							if(count($stepfive)>0){
+							echo '<pre>';print_r($editdata);exit;
+							$editdetails= $this->Hospital_model->update_hospital_details(base64_decode($post['hospital_id']),$editdata);
+							if(count($editdetails)>0){
 								$this->session->set_flashdata('success',"Hospital Details are successfully Saved");
 								redirect('hospital');
 							}else{
